+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Updated README.md with "Command Line Usage", "Upcoming Features" and "Known Issues" sections
+
+### Fixed
+- Fixed type annotations in alkymi's labfile (`lab.py`)
 
 ## [0.0.3] - 2021-01-03
 ### Added