+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Updated README.md with "Command Line Usage", "Upcoming Features" and "Known Issues" sections
-- Included `labfile.py` in static type checking and linting  
+- Included `labfile.py` in static type checking and linting
+- Added a unit test for the built-in `file()` recipe generator
 
 ### Removed
 - "clean" and "clean-cache" commands from Lab (these will be reimplemented properly at a later point)
@@ -17,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed type annotations in alkymi's labfile (`labfile.py`)
+- Fixed reloading of recipe checksums from cache (these needed to be converted from lists to tuples due to JSON
+serialization format). This would cause some recipes to be marked dirty even though they weren't.
 
 ## [0.0.3] - 2021-01-03
 ### Added